Remove views package. Move classes to their respective package.
