+++ app/src/main/java/org/kiwix/kiwixmobile/main/PageBottomTabLayout.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile.views;
+package org.kiwix.kiwixmobile.main;
 
 import android.content.Context;
 import android.support.design.widget.TabLayout;
