+++ app/src/main/java/org/kiwix/kiwixmobile/zim_manager/LanguageSelectDialog.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile.views;
+package org.kiwix.kiwixmobile.zim_manager;
 
 import android.content.Context;
 import android.graphics.Typeface;
